moved displaying favorties list functionality from `mainapp` to `location` and added functionality to only select favorites if the user is a renter
