+++ location/urls.py
@@ -5,4 +5,5 @@ from . import views
 urlpatterns = [
     path("<int:pk>/", views.LocationView.as_view(), name="location"),
     path("<int:pk>/favorite", views.favorites, name="favorite"),
+    path("favlist", views.favlist, name="favlist"),
 ]
