+++ mainapp/templates/nav.html
@@ -49,7 +49,7 @@
                 <ul class="navbar-nav ml-auto">
                     <!-- right aligned items -->
                     <li class="nav-item active">
-                    {% if request.user.is_authenticated %}
+                    {% if request.user.is_authenticated and user.user_type_string == "Renter" %}
                         <a class="nav-link" href="{% url 'favlist' %}">Favorites</a>
                     {% endif %}
                     </li>
