+++ mainapp/views.py
@@ -9,11 +9,3 @@ def index(request):
 @login_required
 def account(request):
     return render(request, "account.html", {"user": request.user})
-
-
-@login_required
-def favlist(request):
-    favorited_apartments = request.user.favorites.all()
-    return render(
-        request, "favlist.html", {"favorited_apartments": favorited_apartments}
-    )
